feat(back-end): delete image after delete

+++ controllers/sauce/deleteSauce.js
@@ -1,19 +1,43 @@
+const fs = require("fs");
+const path = require("path");
+
 const models = {
     Sauce: require("../../models/sauce"),
 };
 
 module.exports = function (req, res) {
     let id = req.params.id;
-    models.Sauce.deleteOne({
+    models.Sauce.findOne({
         _id: id,
-    }).then(function () {
-        res.status(200);
-        res.json({ status: true });
-    })
-        .catch(function () {
+    }).then(function (sauceInfo) {
+        if (sauceInfo) {
+            var splitImage = sauceInfo.imageUrl.split("/");
+            var imagePath = splitImage[splitImage.length - 1];
+
+            fs.unlinkSync(path.join(__dirname, "../../uploads", imagePath));
+
+            models.Sauce.deleteOne({
+                _id: id,
+            }).then(function () {
+                res.status(200);
+                res.json({ status: true });
+            })
+                .catch(function () {
+                    res.status(400);
+                    res.json({ status: true });
+                })
+        }
+        else {
             res.status(404);
             res.json({
                 status: false,
             })
-        });
+            return;
+        }
+    }).catch(function () {
+        res.status(404);
+        res.json({
+            status: false,
+        })
+    });
 }
